refactor #441 : move kvcore unit to tests
