+++ trace/src/lib.rs
@@ -1,5 +1,6 @@
 use std::sync::{Arc, Mutex, Once};
 
+use color_eyre;
 use once_cell::sync::Lazy;
 pub use tracing::{debug, error, info, instrument, trace, warn};
 use tracing_appender::{non_blocking, non_blocking::WorkerGuard, rolling};
