+++ tskv/src/kvcore.rs
@@ -426,283 +426,3 @@ impl KvContext {
         Ok(None)
     }
 }
-
-#[cfg(test)]
-mod test {
-    use std::sync::Arc;
-
-    use chrono::Local;
-    use config::GLOBAL_CONFIG;
-    use futures::{channel::oneshot, future::join_all, SinkExt};
-    use models::{FieldInfo, SeriesInfo, Tag, ValueType};
-    use protos::{
-        kv_service, kv_service::WritePointsRpcResponse, models as fb_models, models_helper,
-    };
-    use serial_test::serial;
-    use snafu::ResultExt;
-    use tokio::sync::{mpsc, oneshot::channel};
-    use trace::{debug, error, info, init_default_global_tracing, warn};
-
-    use crate::{
-        error,
-        kv_option::{TseriesFamDesc, TseriesFamOpt, WalConfig},
-        summary::{Summary, VersionEdit},
-        tseries_family::TimeRange,
-        Error, Task, TsKv,
-    };
-
-    async fn get_tskv() -> TsKv {
-        let opt = crate::kv_option::Options { wal: Arc::new(WalConfig { dir: String::from("/tmp/test/wal"),
-                                                                        ..Default::default() }),
-                                              ..Default::default() };
-
-        TsKv::open(opt).await.unwrap()
-    }
-
-    #[tokio::test]
-    #[serial]
-    async fn test_init() {
-        init_default_global_tracing("tskv_log", "tskv.log", "debug");
-        let tskv = get_tskv().await;
-
-        dbg!("Ok");
-    }
-
-    #[tokio::test]
-    #[serial]
-    async fn test_write() {
-        init_default_global_tracing("tskv_log", "tskv.log", "debug");
-        let tskv = get_tskv().await;
-
-        let database = "db".to_string();
-        let mut fbb = flatbuffers::FlatBufferBuilder::new();
-        let points = models_helper::create_random_points_with_delta(&mut fbb, 1);
-        fbb.finish(points, None);
-        let points = fbb.finished_data().to_vec();
-        let request = kv_service::WritePointsRpcRequest { version: 1, database, points };
-
-        tskv.write(request).await.unwrap();
-    }
-
-    // remove repeat sid and fields_id
-    pub fn remove_duplicates(nums: &mut [u64]) -> usize {
-        init_default_global_tracing("tskv_log", "tskv.log", "debug");
-        if nums.len() <= 1 {
-            return nums.len() as usize;
-        }
-        let mut l = 1;
-        for r in 1..nums.len() {
-            if nums[r] != nums[l - 1] {
-                nums[l] = nums[r];
-                l += 1;
-            }
-        }
-        l as usize
-    }
-
-    // tips : to test all read method, we can use a small MAX_MEMCACHE_SIZE
-    #[tokio::test]
-    #[serial]
-    async fn test_read() {
-        init_default_global_tracing("tskv_log", "tskv.log", "debug");
-        let tskv = get_tskv().await;
-
-        let database = "db".to_string();
-        let mut fbb = flatbuffers::FlatBufferBuilder::new();
-        let points = models_helper::create_random_points_with_delta(&mut fbb, 20);
-        fbb.finish(points, None);
-        let points = fbb.finished_data().to_vec();
-        let request = kv_service::WritePointsRpcRequest { version: 1, database, points };
-
-        tskv.write(request.clone()).await.unwrap();
-
-        let shared_write_batch = Arc::new(request.points);
-        let fb_points = flatbuffers::root::<fb_models::Points>(&shared_write_batch).context(error::InvalidFlatbufferSnafu).unwrap();
-        let mut sids = vec![];
-        let mut fields_id = vec![];
-        for point in fb_points.points().unwrap() {
-            let mut info =
-                SeriesInfo::from_flatbuffers(&point).context(error::InvalidModelSnafu).unwrap();
-            info.finish();
-            sids.push(info.series_id());
-            for field in info.field_infos().iter() {
-                fields_id.push(field.field_id());
-            }
-        }
-
-        sids.sort_unstable();
-        fields_id.sort_unstable();
-        let l = remove_duplicates(&mut sids);
-        sids = sids[0..l].to_owned();
-        let l = remove_duplicates(&mut fields_id);
-        fields_id = fields_id[0..l].to_owned();
-        tskv.read(sids, &TimeRange::new(Local::now().timestamp_millis() + 100, 0), fields_id).await;
-    }
-
-    #[tokio::test]
-    #[serial]
-    #[ignore]
-    async fn test_delete_cache() {
-        init_default_global_tracing("tskv_log", "tskv.log", "debug");
-        let tskv = get_tskv().await;
-        let database = "db".to_string();
-        let mut fbb = flatbuffers::FlatBufferBuilder::new();
-        let points = models_helper::create_random_points_with_delta(&mut fbb, 5);
-        fbb.finish(points, None);
-        let points = fbb.finished_data().to_vec();
-        let request = kv_service::WritePointsRpcRequest { version: 1, database, points };
-
-        tskv.write(request.clone()).await.unwrap();
-
-        let shared_write_batch = Arc::new(request.points);
-        let fb_points = flatbuffers::root::<fb_models::Points>(&shared_write_batch).context(error::InvalidFlatbufferSnafu).unwrap();
-        let mut sids = vec![];
-        let mut fields_id = vec![];
-        for point in fb_points.points().unwrap() {
-            let mut info =
-                SeriesInfo::from_flatbuffers(&point).context(error::InvalidModelSnafu).unwrap();
-            info.finish();
-            sids.push(info.series_id());
-            for field in info.field_infos().iter() {
-                fields_id.push(field.field_id());
-            }
-        }
-
-        sids.sort_unstable();
-        fields_id.sort_unstable();
-        let l = remove_duplicates(&mut sids);
-        sids = sids[0..l].to_owned();
-        let l = remove_duplicates(&mut fields_id);
-        fields_id = fields_id[0..l].to_owned();
-        tskv.read(sids.clone(),
-                  &TimeRange::new(Local::now().timestamp_millis() + 100, 0),
-                  fields_id.clone())
-            .await;
-        info!("delete delta data");
-        tskv.delete_series(sids.clone(), 1, 1).await.unwrap();
-        tskv.read(sids.clone(),
-                  &TimeRange::new(Local::now().timestamp_millis() + 100, 0),
-                  fields_id.clone())
-            .await;
-    }
-
-    #[tokio::test]
-    #[serial]
-    #[ignore]
-    async fn test_big_write() {
-        init_default_global_tracing("tskv_log", "tskv.log", "debug");
-        let tskv = get_tskv().await;
-
-        for i in 0..100 {
-            let database = "db".to_string();
-            let mut fbb = flatbuffers::FlatBufferBuilder::new();
-            let points = models_helper::create_big_random_points(&mut fbb, 10);
-            fbb.finish(points, None);
-            let points = fbb.finished_data().to_vec();
-
-            let request = kv_service::WritePointsRpcRequest { version: 1, database, points };
-
-            tskv.write(request).await.unwrap();
-        }
-    }
-
-    #[tokio::test]
-    #[serial]
-    async fn test_insert_cache() {
-        init_default_global_tracing("tskv_log", "tskv.log", "debug");
-        let tskv = get_tskv().await;
-
-        let database = "db".to_string();
-        let mut fbb = flatbuffers::FlatBufferBuilder::new();
-        let points = models_helper::create_random_points_with_delta(&mut fbb, 1);
-        fbb.finish(points, None);
-        let points = fbb.finished_data();
-
-        tskv.insert_cache(1, points).await.unwrap();
-    }
-
-    #[tokio::test]
-    #[serial]
-    async fn test_start() {
-        init_default_global_tracing("tskv_log", "tskv.log", "debug");
-        let tskv = get_tskv().await;
-
-        let (wal_sender, wal_receiver) = mpsc::unbounded_channel();
-        let (tx, rx) = channel();
-
-        let database = "db".to_string();
-        let mut fbb = flatbuffers::FlatBufferBuilder::new();
-        let points = models_helper::create_random_points_with_delta(&mut fbb, 1);
-        fbb.finish(points, None);
-        let points = fbb.finished_data().to_vec();
-        let req = kv_service::WritePointsRpcRequest { version: 1, database, points };
-
-        wal_sender.send(Task::WritePoints { req, tx }).unwrap();
-
-        TsKv::start(tskv, wal_receiver);
-
-        match rx.await {
-            Ok(Ok(resp)) => info!("successful"),
-            _ => panic!("wrong"),
-        };
-    }
-
-    #[tokio::test]
-    #[serial]
-    async fn test_add_del_tsf() {
-        init_default_global_tracing("tskv_log", "tskv.log", "debug");
-        let tskv = get_tskv().await;
-        let opt = crate::kv_option::Options { wal: Arc::new(WalConfig { dir: String::from("/tmp/test/wal"),
-                                                                        ..Default::default() }),
-                                              ..Default::default() };
-        let shared_options = Arc::new(opt);
-        let summary = Summary::new(&shared_options.db).await.unwrap();
-        summary.global_context().next_tsf_id();
-        let tf_id = summary.global_context().max_tsf_id();
-
-        tskv.version_set()
-            .write()
-            .add_tsfamily(tf_id,
-                          "hello".to_string(),
-                          0,
-                          0,
-                          Arc::new(TseriesFamOpt::default()),
-                          tskv.summary_task_sender.clone())
-            .await;
-        assert_eq!(tskv.version_set().read().tsf_num(), (GLOBAL_CONFIG.tsfamily_num + 1) as usize);
-
-        tskv.version_set()
-            .write()
-            .del_tsfamily(tf_id, "hello".to_string(), tskv.summary_task_sender.clone());
-        assert_eq!(tskv.version_set().read().tsf_num(), GLOBAL_CONFIG.tsfamily_num as usize);
-
-        info!("success");
-    }
-
-    #[tokio::test]
-    #[serial]
-    async fn test_flush_delta() {
-        init_default_global_tracing("tskv_log", "tskv.log", "debug");
-        let tskv = get_tskv().await;
-        let database = "db".to_string();
-        let mut fbb = flatbuffers::FlatBufferBuilder::new();
-        let points = models_helper::create_random_points_include_delta(&mut fbb, 20);
-        fbb.finish(points, None);
-        let points = fbb.finished_data().to_vec();
-        let request = kv_service::WritePointsRpcRequest { version: 1, database, points };
-
-        tskv.write(request).await.unwrap();
-    }
-
-    #[tokio::test]
-    #[serial]
-    async fn test_log() {
-        init_default_global_tracing("tskv_log", "tskv.log", "debug");
-        let tskv = get_tskv().await;
-        info!("hello");
-        warn!("hello");
-        debug!("hello");
-        error!("hello"); //maybe we can use panic directly
-        // panic!("hello");
-    }
-}
